+++ Tests/SentryTests/HybridSDK/SentryInternalScopeApiTests.swift
@@ -18,8 +18,9 @@ final class SentryInternalScopeApiTests: XCTestCase {
         XCTAssertNotNil(contexts["trace"])
     }
 
-    private struct Dependencies: HubProvider {
+    private struct Dependencies: HubProvider, CurrentScopeStorageProvider {
         let hub: Hub
+        let currentScopeStorage = SentryCurrentScopeStorage()
 
         init(scope: Scope) {
             hub = TestHub(scope: scope)
@@ -40,6 +41,10 @@ final class SentryInternalScopeApiTests: XCTestCase {
         func storeEnvelope(_ envelope: SentryEnvelope) {}
 
         func captureEnvelope(_ envelope: SentryEnvelope) {}
+
+        func captureNonTerminatingEnvelope(_ envelope: SentryEnvelope) {}
+
+        func updateSessionForDroppedEventNonTerminating(unhandled: Bool) {}
 
         func captureErrorEvent(event: Event) {}